Move ExhibitionHeader to its own directory
