+++ common/views/components/ExhibitionHeader/ExhibitionHeader.js
