+++ common/views/components/ExhibitionHeader/README.md
@@ -1,0 +1,3 @@
+## Exhibition header
+
+TODO: add purpose of exhibition header.
